feat(messages): NICK messages implemented
