+++ includes/utils/Utils.hpp
@@ -3,6 +3,7 @@
 
 # include <iostream>
 # include <vector>
+#include <map>
 # include <string>
 # include "../client/Client.hpp"
 
@@ -25,7 +26,7 @@ const std::string	MSG_NEEDMOREPARAMS = ":Not enough parameters";
 
 
 std::string	authPass(Client &client, std::string password, std::string server_password);
-void	setClientNick(std::string nickname, Client &client);
+std::string	setClientNick(std::string nickname, Client &client, std::map<int, Client> &clients);
 bool	setClientUsername(std::string message, std::vector<std::string> split_msg, Client &client);
 
 #endif
