+++ src/commands/NICK.cpp
@@ -10,31 +10,43 @@ bool	isSpecial(char c)
 	return (false);
 }
 
-void	setClientNick(std::string nickname, Client &client)
+bool	isNewNick(std::map<int, Client> &clients, std::string nickname)
 {
-	if (client.getIsAuthenticated())
+	for (std::map<int, Client>::iterator it = clients.begin(); it != clients.end(); ++it)
 	{
-		if (nickname.length() < 1 || nickname.length() > 9)
-		{
-			std::cout << "Ivalid nick: nickname must be between 1 and 9" << std::endl;
-			return ;
-		}
-		if (!std::isalpha(nickname[0]) && !isSpecial(nickname[0]))
-		{
-			std::cout << "Ivalid nick: first character must be alpha or special" << std::endl;
-			return ;
-		}
-		for (std::string::iterator it = nickname.begin(); it != nickname.end(); ++it)
-		{
-			if (!std::isalpha(*it) && !std::isdigit(*it) && !isSpecial(*it))
-			{
-				std::cout << "Ivalid nick: invalid character: " << *it << std::endl;
-				return ;
-			}
-		}
-		client.setNickname(nickname);
-		std::cout << "nick set: " << client.getNickname() << std::endl;
+		Client client = it->second;
+		if (client.getNickname().compare(nickname) == 0)
+			return (false);
 	}
-	else
-		std::cout << "Client not authenticated" << std::endl;
+	return (true);
+}
+
+std::string	setClientNick(std::string nickname, Client &client, std::map<int, Client> &clients)
+{
+	if (nickname.length() < 1 || nickname.length() > 9)
+	{
+		std::cout << "Ivalid nick: nickname must be between 1 and 9. Client FD: " << client.getFd() << std::endl;
+		return (ERR_ERRONEUSNICKNAME);
+	}
+	if (!std::isalpha(nickname[0]) && !isSpecial(nickname[0]))
+	{
+		std::cout << "Ivalid nick: first character must be alpha or special. Client FD: " << client.getFd() << std::endl;
+		return (ERR_ERRONEUSNICKNAME);
+	}
+	if (!isNewNick(clients, nickname))
+	{
+		std::cout << "Nick " << nickname << " has already been registered. Client FD: " << client.getFd() << std::endl;
+		return (ERR_NICKNAMEINUSE);
+	}
+	for (std::string::iterator it = nickname.begin(); it != nickname.end(); ++it)
+	{
+		if (!std::isalpha(*it) && !std::isdigit(*it) && !isSpecial(*it))
+		{
+			std::cout << "Ivalid nick: invalid character: " << *it << ". Client FD: " << client.getFd() << std::endl;
+			return (ERR_ERRONEUSNICKNAME);
+		}
+	}
+	client.setNickname(nickname);
+	std::cout << "nick set: " << client.getNickname() << std::endl;
+	return (RPL_SUCCESS);
 }
